requirements.txt
asgiref==3.3.0
chardet==3.0.4
coverage==5.2
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.3.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
sendgrid==6.9.3
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==2.0.3
urllib3==1.25.10
numpy
dlib
face_recognition
imutils
opencv-python
Pillow~=9.2.0